# ext/bcmath: Improving mul performance 2